Migrates to Babel 6.x
Babel 6.x uses presets and plugins that need to be explicitly specified
and installed.

To conserve behavior most closely to prior to this change, the 'env' preset
would be used as the default preset if none is specified by the user.
However, that would in turn require the user to install this preset on
their own; a dependency in this package does not suffice.

Hence, there is no default preset, and if none are specified, the compiler
simply reproduces the code it receives as input. This is in fact a valid
use case, another reason not to set any default preset.